// server/src/modules/hotel/hotel.interface.ts
export interface IHotelFilter {
  location?: string;
  searchTerm?: string;
}

export interface ICreateHotelInput {
  name: string;
  location: string;
  description: string;
  pricePerNight: number;
  coverImage: string;
  images?: string[];
  amenities?: string[];
  contactPhone?: string;
}

export type IUpdateHotelInput = Partial<ICreateHotelInput>;